# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Documentation API Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://connect.test",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "getPostsList",
                "requestBody": {
                    "$ref": "#/components/schemas/Post"
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "description": "Creates a new post and returns it",
                "operationId": "storePost",
                "requestBody": {
                    "description": "Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post by ID",
                "description": "Returns a single post",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update an existing post",
                "operationId": "dcf5b8dcbaee8bb9fa6febfd2eeb0589",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "description": "Deletes a post",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PostRequest": {
                "title": "Post request",
                "description": "Store Post request body data",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the post",
                        "type": "string",
                        "example": "Post title"
                    },
                    "content": {
                        "description": "Content of the post",
                        "type": "string",
                        "example": "Post content"
                    }
                },
                "type": "object"
            },
            "PostResource": {
                "title": "Post resource",
                "description": "Post resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/Post_id"
                            },
                            "title": {
                                "$ref": "#/components/schemas/Post_title"
                            },
                            "content": {
                                "$ref": "#/components/schemas/Post_content"
                            },
                            "status": {
                                "$ref": "#/components/schemas/Post_status"
                            },
                            "author_by": {
                                "$ref": "#/components/schemas/Post_author_by"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PostResource"
                }
            },
            "Post_id": {
                "title": "Id",
                "description": "ID",
                "format": "int64",
                "default": 1
            },
            "Post_title": {
                "title": "Title",
                "description": "Title",
                "format": "string",
                "default": "Title blog"
            },
            "Post_content": {
                "title": "Content",
                "description": "Content of Blog",
                "format": "string",
                "default": "Content blog"
            },
            "Post_status": {
                "title": "Status",
                "description": "Status of Blog",
                "format": "boolean",
                "default": "0"
            },
            "Post_author_id": {
                "title": "Author",
                "description": "Author of Blog",
                "format": "integer",
                "default": "1"
            },
            "Post_author_name": {
                "title": "Author",
                "description": "Author of Blog",
                "format": "string",
                "default": "Author Spain"
            },
            "Post_author_by": {
                "title": "Author",
                "description": "Author of Blog",
                "format": "string",
                "default": "Author Spain"
            },
            "Post": {
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Post_id"
                    },
                    "title": {
                        "$ref": "#/components/schemas/Post_title"
                    },
                    "content": {
                        "$ref": "#/components/schemas/Post_content"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Post_status"
                    },
                    "author_id": {
                        "$ref": "#/components/schemas/Post_author_id"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}